feat: add workspace support to go-mod-tidy task
