+++ .pre-commit-hooks.yaml
@@ -83,7 +83,7 @@
     description: "Runs `go generate`, requires golang"
 -   id: go-mod-tidy
     name: 'go-mod-tidy'
-    entry: run-go-mod-tidy.sh
+    entry: run-go-mod-tidy.py
     pass_filenames: false
     language: 'script'
     description: "Runs `go mod tidy -v`, requires golang"
